Clarify replica module boundaries
