+++ src/ogcat/replica_types.py
@@ -1,0 +1,13 @@
+"""Shared replica type aliases."""
+
+from __future__ import annotations
+
+from typing import Literal
+
+ReplicaMode = Literal["symlink"]
+ReplicaRole = Literal["template_link", "view_link"]
+
+__all__ = [
+    "ReplicaMode",
+    "ReplicaRole",
+]
